BETA BUILD 1

+++ src/core/idt.h
@@ -4,6 +4,6 @@
 #include "common.h"
 
 void idt_init(void);
-void idt_set_gate(uint8_t num, uint32_t base, uint16_t sel, uint8_t flags);
+void idt_set_gate(uint8_t num, uint64_t base, uint16_t sel, uint8_t flags);
 
 #endif
